Create actions_for_climatechange_examples
For the cotigecy plan actions I used an example of a list, pretty simple to enumerate tasks for self-reliance, for the real app it will be probably a dictionary. I then used an example of a dictionary using boolean for status For actions to reduce climate change I used for loops and tuples for estimated CO2 savings. I used a tuple because each item had two or more attributes and the for loop for unpacking.