Create social_media_blog_examples
I used an example of a list of dictionaries for storing user posts and for loop for processing and displaying each item. Then I used a while loop to approve all posts until I find one that is flagged. I also did an example for my blog feature using a for loop with enumerate for numbered display.