Create num_cycles_shampoo_instructions
I do need to do these a few more times alternating between ranges and iterable lists because I immediately wanted to do 'for N in num_cycles', but when it is an integer I need to do a range and N += 1 is unnecessary because the range does the adding 1. This is just more practice for functions with a loop variable.