Create counting_characters
This is a lab testing on my ability to write a code that allows the user to give a character followed by a word or phrase and then it counts how many of those characters are in the word and it is case sensitive. At first, because I was learning split and join I thought to split the character and phrase but then I realized in this case it is simpler to use indexing of the user input, char is user_input[0] and the phrase is everything after the 1 index. and then it is counting using count and give the output