Create Name_vs_integr_ErrorHandling
The given program reads a list of single-word first names and ages (ending with -1), and outputs that list with the age incremented. The program fails and throws an exception if the second input on a line is a string rather than an integer. Add try and except blocks to catch the ValueError exception and output 0 for the age.